requirements.txt
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.6
pytest==8.2.2
python-dotenv==1.0.1
django-cors-headers==4.4.0

